Reduce the set of tests required to reproduce flaky behavior.
DynamicObjectIntegrationTest only fails when run after
ScriptPluginClassLoadingIntegrationTest AND DynamicObjectIntegrationTest

Test execution order is _host_ dependent, and the right ordering is
exemplified by ubuntu28 and ubuntu27 agents.

Removing the ConfigureLogging @Rule from AbstractFileWatcherTest fixes
flaky behavior, so this is related to the failure.

ScriptPluginClassLoadingIntegrationTest is related, but unknown how yet
other than we know that no-op tests _do not_ cause the flaky behavior.
So something in the 1 remaining test is involved.