Use `Binary` in place of `BinarySpec` in more places
This change removes most of the dependency of library resolution on
`BinarySpec`, which should allow things like `PrebuiltLibrary` to participate
in dependency resolution.

It's a bit/lot of a hack, but it takes us in the right direction, where the "spec"
part of `BinarySpec` is about how it's built, not how it's consumed.

The JVM implementation is still casting to BinarySpec quite a lot.