Rework annotated task property handling
This change postpones adding annotated inputs and outputs to the task's `TaskInputs` and `TaskOutputs` until the execution plan is being made. The advantage of this is that we don't need to defer the evaluation of the properties via `Callable`s, and thus can process the actual values.

This way we can later handle properties that contain multiple values based on their actual runtime values. In particular this enables adding support for properties like:

    @Nested
    Iterable<MyData> myData