Fix failing tests by passing the `Configuration` context to the dependency set
This commit fixes the "non fluid" dependencies use case by introducing a way to pass the consumer context
to the dependency resolution engine when constructing task dependencies. This is done by adding the
`Configuration` instance to `DefaultDependencySet`, and by adding a method to `ProjectDependencyInternal`
so that we can use the attributes to choose the appropriate configuration.