Duplicate the selection logic in `LocalComponentDependencyMetadata#selectConfigurations`
This commit duplicates the logic of selection of configurations in `LocalComponentDependencyMetadata#selectConfigurations` in order to make sure
that when a configuration inherits from another configuration, when the project dependency is defined on the parent configuration, we still select
the right configuration based on the client requirements.

The implementation is a bit unsatisfactory in the sense that it does duplicate logic that exists in `ProjectDependency`, but for metadata instead
of dependencies themselves. It also triggers the realization of actions for empty configurations on all the configurations of the target project
(which may or may not be an issue), instead of just the selected configuration. This is because we cannot defer it, since when we know we're going
to use a configuration, it's already too late.

It's worth mentioning that the `dependencyConfiguration` used in `LocalComponentDependencyMetadata` reflects the dependency configuration which is
*declared* in the dependency notation (with the bias that if the configuration is not declared, it will fallback to _default_, which is going to
be addressed in a future commit). It means that we should avoid calling this method for anything related to dependency resolution itself, until
we're able to tell whether the configuration was specified by the user or not.